fix: CORUI-6264: One component per file InputGroup (#426)
* Rename file

* Rename temp

* Restore original

* Separate component files

* Rename file

* Rename temp

* Restore original

* Separate test files
