+++ src/index.js
@@ -27,7 +27,8 @@ export { Icon } from './Icon/Icon';
 export { Identifier } from './Identifier/Identifier';
 export { Image } from './Image/Image';
 export { InlineHelp } from './InlineHelp/InlineHelp';
-export { InputGroup, FormGroup } from './InputGroup/InputGroup';
+export { default as InputGroup } from './InputGroup/InputGroup';
+export { default as FormGroup } from './InputGroup/FormGroup';
 export { ListGroup, ListGroupItem, ListGroupItemActions, ListGroupItemCheckbox } from './ListGroup/ListGroup';
 export { LocalizationEditor } from './LocalizationEditor/LocalizationEditor';
 export { Menu, MenuList, MenuItem, MenuGroup } from './Menu/Menu';
